+++ packages/uptrace-core/src/index.ts
@@ -1,4 +1,2 @@
 export * from './config'
-export * from './client'
-export * from './types'
 export * from './version'