+++ packages/uptrace-web/src/sdk.ts
@@ -1,8 +1,14 @@
-import { Span } from '@opentelemetry/api'
+import { trace, context, Span, Attributes } from '@opentelemetry/api'
+import { logs, Logger, SeverityNumber } from '@opentelemetry/api-logs'
 import { WebTracerProvider } from '@opentelemetry/sdk-trace-web'
 import { MeterProvider } from '@opentelemetry/sdk-metrics'
 import { LoggerProvider } from '@opentelemetry/sdk-logs'
 import { registerInstrumentations } from '@opentelemetry/instrumentation'
+import {
+  ATTR_EXCEPTION_MESSAGE,
+  ATTR_EXCEPTION_STACKTRACE,
+  ATTR_EXCEPTION_TYPE,
+} from '@opentelemetry/semantic-conventions'
 
 import { createResource, parseDsn, Dsn } from '@uptrace/core'
 import { initConfig, Config } from './config'
@@ -10,10 +16,12 @@ import { browserAttributes, entryPageAttributes } from './resources'
 import { configureTracing } from './tracing'
 import { configureMetrics } from './metrics'
 import { configureLogs } from './logs'
+import { VERSION } from './version'
 
 export class WebSDK {
   private _conf: Config
   private _dsn: Dsn
+  private _logger: Logger
 
   private _tracerProvider?: WebTracerProvider
   private _meterProvider?: MeterProvider
@@ -23,6 +31,7 @@ export class WebSDK {
     initConfig(conf)
     this._conf = conf
     this._dsn = parseDsn(conf.dsn)
+    this._logger = logs.getLogger('uptrace-js', VERSION)
   }
 
   public start(): void {
@@ -81,6 +90,41 @@ export class WebSDK {
     const traceId = ctx?.traceId ?? '<no trace>'
     const spanId = ctx?.spanId ?? '<no span>'
     return `${this._dsn.siteUrl()}/traces/${traceId}?span_id=${spanId}`
+  }
+
+  public reportException(err: Error | string, attributes: Attributes = {}) {
+    const span = trace.getSpan(context.active())
+    if (span) {
+      if (typeof err === 'string') {
+        attributes[ATTR_EXCEPTION_MESSAGE] = err
+        span.addEvent('exception', attributes)
+      } else {
+        span.recordException(err)
+      }
+      return
+    }
+
+    if (typeof err === 'string') {
+      this._logger.emit({
+        severityNumber: SeverityNumber.ERROR,
+        severityText: 'ERROR',
+        body: err,
+        attributes,
+      })
+      return
+    }
+
+    this._logger.emit({
+      severityNumber: SeverityNumber.ERROR,
+      severityText: 'ERROR',
+      body: err.message,
+      attributes: {
+        ...attributes,
+        [ATTR_EXCEPTION_TYPE]: err.name,
+        [ATTR_EXCEPTION_MESSAGE]: err.message,
+        [ATTR_EXCEPTION_STACKTRACE]: err.stack,
+      },
+    })
   }
 }